api/requirements.txt
#Lib Tests
model-mommy==1.6.0
coverage==4.5.4

#Lib Django
gunicorn==19.8.1
Django==2.2.4
django-currentuser==0.4.2
django-cors-headers==3.1.0
django-extensions==2.2.1
django-filter==2.2.0
djangorestframework==3.10.2
requests==2.22.0

#Lib Utils
freeze==1.0.10
python-dotenv==0.10.3
python-dateutil==2.8.0

#database
#TODO: Docker remeber tip: before install psycopg: sudo apt-get install libpq-dev
psycopg2==2.7.3.1
